backend/requirements.txt
fastapi
uvicorn
replicate
python-dotenv
pydantic-settings
pydantic
typing-extensions
google-generativeai
google-genai
redis
